requirements.txt
# Python 3.10 or higher

python-telegram-bot >= 20.0a4
ruamel.yaml
